# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Garage API",
        "description": "Documentation de l'API Gestion Garage",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Lister les clients",
                "description": "Display a listing of the resource.",
                "operationId": "a53b083ee60f942c405b55be2f2117b9",
                "responses": {
                    "200": {
                        "description": "Liste des clients recuperee avec succes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Creer un client",
                "description": "Store a newly created resource in storage.",
                "operationId": "77bf6aecca49f5534d3970adeae37116",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Diallo"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Mamadou"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "622000000"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "Conakry"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client cree avec succes"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/clients/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Afficher un client",
                "description": "Display the specified resource.",
                "operationId": "44968021d7bb28d80a4750383aaba04d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client recupere avec succes"
                    },
                    "404": {
                        "description": "Client introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Modifier un client",
                "description": "Update the specified resource in storage.",
                "operationId": "6e00beceb6af929a7b1fb096ad218c8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Bah"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Aissatou"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "628111111"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "Kipe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client modifie avec succes"
                    },
                    "404": {
                        "description": "Client introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Supprimer un client",
                "description": "Remove the specified resource from storage.",
                "operationId": "d2745e4ba1c3e0fd19d77300ea991354",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client supprime avec succes"
                    },
                    "404": {
                        "description": "Client introuvable"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Clients"
                ],
                "summary": "Modifier partiellement un client",
                "operationId": "ad7ced70583f934a192fee0fd0621a36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Bah"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Aissatou"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "628111111"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "Kipe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client modifie avec succes"
                    },
                    "404": {
                        "description": "Client introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
    },
...
}